requirements.txt
apache-airflow==2.10.5
transformers
torch
torchvision
pytesseract
pillow
pandas>=1.3.0
opencv-python		
setuptools
wheel
scipy
bitsandbytes
accelerate
